Move internal/config to config
