+++ cmd/gazelle/BUILD.bazel
@@ -15,7 +15,7 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/cmd/gazelle",
     visibility = ["//visibility:private"],
     deps = [
-        "//internal/config:go_default_library",
+        "//config:go_default_library",
         "//internal/flag:go_default_library",
         "//internal/language:go_default_library",
         "//internal/language/go:go_default_library",
@@ -62,7 +62,7 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//internal/config:go_default_library",
+        "//config:go_default_library",
         "//internal/testtools:go_default_library",
         "//internal/wspace:go_default_library",
     ],
