+++ internal/language/lang.go
@@ -16,7 +16,7 @@ limitations under the License.
 package language
 
 import (
-	"github.com/bazelbuild/bazel-gazelle/internal/config"
+	"github.com/bazelbuild/bazel-gazelle/config"
 	"github.com/bazelbuild/bazel-gazelle/internal/resolve"
 	"github.com/bazelbuild/bazel-gazelle/rule"
 )
