+++ internal/language/proto/BUILD.bazel
@@ -27,7 +27,7 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/internal/language/proto",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/config:go_default_library",
+        "//config:go_default_library",
         "//internal/language:go_default_library",
         "//internal/repos:go_default_library",
         "//internal/resolve:go_default_library",
@@ -46,7 +46,7 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
     deps = [
-        "//internal/config:go_default_library",
+        "//config:go_default_library",
         "//internal/language:go_default_library",
         "//internal/merger:go_default_library",
         "//internal/repos:go_default_library",
