+++ internal/language/proto/config.go
@@ -21,7 +21,7 @@ import (
 	"log"
 	"path"
 
-	"github.com/bazelbuild/bazel-gazelle/internal/config"
+	"github.com/bazelbuild/bazel-gazelle/config"
 	"github.com/bazelbuild/bazel-gazelle/rule"
 )
