+++ internal/resolve/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/internal/resolve",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/config:go_default_library",
+        "//config:go_default_library",
         "//internal/repos:go_default_library",
         "//label:go_default_library",
         "//rule:go_default_library",
