+++ internal/testtools/config.go
@@ -19,7 +19,7 @@ import (
 	"flag"
 	"testing"
 
-	"github.com/bazelbuild/bazel-gazelle/internal/config"
+	"github.com/bazelbuild/bazel-gazelle/config"
 	"github.com/bazelbuild/bazel-gazelle/internal/language"
 )
