+++ internal/walk/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/internal/walk",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/config:go_default_library",
+        "//config:go_default_library",
         "//internal/flag:go_default_library",
         "//pathtools:go_default_library",
         "//rule:go_default_library",
@@ -21,7 +21,7 @@ go_test(
     srcs = ["walk_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//internal/config:go_default_library",
+        "//config:go_default_library",
         "//internal/testtools:go_default_library",
         "//rule:go_default_library",
     ],
